+++ Example09/src/Application.cpp
@@ -128,7 +128,7 @@ bool Application::applicationInit()
 	descriptorPoolCreateInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
 	descriptorPoolCreateInfo.poolSizeCount = 2;
 	descriptorPoolCreateInfo.pPoolSizes = descriptorPoolSizes;
-	descriptorPoolCreateInfo.maxSets = commandBuffers.size();
+	descriptorPoolCreateInfo.maxSets = static_cast<uint32_t>(commandBuffers.size());
 
 	result = vkCreateDescriptorPool(device, &descriptorPoolCreateInfo, nullptr, &descriptorPool);
 	if (result != VK_SUCCESS)
